Fix various bugs

+++ game/client/c_playerresource.h
@@ -83,7 +83,7 @@ public : // IGameResources intreface
 	virtual int		GetTeamPosType( int iIndex );
 	virtual int		GetTeamPosIndex( int iIndex );
 	virtual int		GetTeamToJoin( int iIndex );
-	virtual float	GetNextJoin( int iIndex );
+	virtual int		GetNextJoin( int iIndex );
 	virtual bool	IsCardBanned( int iIndex );
 
 	virtual void ClientThink();
@@ -128,7 +128,7 @@ protected:
 	int		m_TeamPosNum[MAX_PLAYERS+1];
 
 	int		m_TeamToJoin[MAX_PLAYERS+1];
-	float	m_NextJoin[MAX_PLAYERS+1];
+	int		m_NextJoin[MAX_PLAYERS+1];
 	bool	m_IsCardBanned[MAX_PLAYERS+1];
 };
